+++ miniRT/srcs/collisions.c
@@ -17,10 +17,10 @@ https://hugi.scene.org/online/hugi24/coding%20graphics%20chris%2\
 	  0dragan%20raytracing%20shapes.htm
 */
 
-float	quadratic_form(float a, float b, float c)
+double	quadratic_form(double a, double b, double c)
 {
-	float	t_plus;
-	float	t_minu;
+	double	t_plus;
+	double	t_minu;
 
 	t_plus = (-1 * b + sqrt(b * b - 4 * a * c)) / (2 * a);
 	t_minu = (-1 * b - sqrt(b * b - 4 * a * c)) / (2 * a);
@@ -35,8 +35,8 @@ float	quadratic_form(float a, float b, float c)
 bool	sphere_collision(t_sphere *sp, t_ray *r)
 {
 	t_vector	x;
-	float		c;
-	float		t;
+	double		c;
+	double		t;
 
 	x = vector_create(r->p0, sp->pos);
 	c = vector_dot(x, x) - sp->d_squared;
@@ -64,9 +64,9 @@ void	bases_aux_2(t_ray *r, t_vector v)
 		r->norm_v = vector_mult_const(v, -1);
 }
 
-float	bases_aux(float *n, t_pos *pos, t_vector *vec, t_cylinder *c)
+double	bases_aux(double *n, t_pos *pos, t_vector *vec, t_cylinder *c)
 {
-	float	t[3];
+	double	t[3];
 
 	if (n[1] == 0)
 		t[1] = -1;
@@ -93,9 +93,9 @@ float	bases_aux(float *n, t_pos *pos, t_vector *vec, t_cylinder *c)
 	return (t[0]);
 }
 
-float	bases(t_cylinder *c, t_ray *r, float t)
+double	bases(t_cylinder *c, t_ray *r, double t)
 {
-	float		n[4];
+	double		n[4];
 	t_pos		pos;
 	t_vector	vec;
