+++ miniRT/srcs/key_hook.c
@@ -33,11 +33,11 @@ void	key_hook_aux_3(int keycode, t_mini *m)
 	else if (keycode == Z && m->action == 'r')
 		dim_rot_trans(m, 0, 0, 'z');
 	else if (keycode == X && !m->hl)
-		cam_rotation(m, 'x', (float [2]){0});
+		cam_rotation(m, 'x', (double [2]){0});
 	else if (keycode == Y && !m->hl)
-		cam_rotation(m, 'y', (float [2]){0});
+		cam_rotation(m, 'y', (double [2]){0});
 	else if (keycode == Z && !m->hl)
-		cam_rotation(m, 'z', (float [2]){0});
+		cam_rotation(m, 'z', (double [2]){0});
 }
 
 void	key_hook_aux_2(int keycode, t_mini *m)
